requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5
openpyxl>=3.0.9
xlrd>=2.0.1
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0
